+++ src/main/kotlin/cloud/glitchdev/rfu/gui/UIScheme.kt
@@ -1,4 +1,4 @@
-﻿package cloud.glitchdev.rfu.gui
+package cloud.glitchdev.rfu.gui
 
 import cloud.glitchdev.rfu.constants.fishing.FishingIslands
 import java.awt.Color
@@ -111,6 +111,8 @@ object UIScheme {
     val pfConditionCardEnderman9Border = Color(170, 0, 170)
     val pfConditionCardBrainFood = Color(128, 42, 128)
     val pfConditionCardBrainFoodBorder = Color(255, 85, 255)
+    val pfConditionCardBloodshot = Color(100, 20, 20)
+    val pfConditionCardBloodshotBorder = veryHardDifficultyColor
     val pfConditionCardCanJoin = Color(11, 104, 25)
     val pfConditionCardCanJoinBorder = Color(23, 209, 51)
     val pfConditionCardIsland = Color(45, 45, 45)