+++ src/main/kotlin/cloud/glitchdev/rfu/gui/components/partyfinder/UIToggleCard.kt
@@ -49,6 +49,7 @@ class UIToggleCard(
                 "looting_5" -> UIScheme.pfConditionCardLooting5
                 "enderman_9" -> UIScheme.pfConditionCardEnderman9
                 "brain_food" -> UIScheme.pfConditionCardBrainFood
+                "bloodshot" -> UIScheme.pfConditionCardBloodshot
                 "can_join" -> UIScheme.pfConditionCardCanJoin
                 "location" -> UIScheme.getIslandColor(option.label)
                 else -> UIScheme.pfConditionCardUnknown
@@ -65,6 +66,7 @@ class UIToggleCard(
                 "looting_5" -> UIScheme.pfConditionCardLooting5Border
                 "enderman_9" -> UIScheme.pfConditionCardEnderman9Border
                 "brain_food" -> UIScheme.pfConditionCardBrainFoodBorder
+                "bloodshot" -> UIScheme.pfConditionCardBloodshotBorder
                 "can_join" -> UIScheme.pfConditionCardCanJoinBorder
                 "location" -> UIScheme.getIslandBorderColor(option.label)
                 else -> UIScheme.pfConditionCardUnknownBorder
@@ -87,6 +89,7 @@ class UIToggleCard(
                 "looting_5" -> "/assets/rfu/ui/looting_5.png"
                 "enderman_9" -> "/assets/rfu/ui/enderman_9.png"
                 "brain_food" -> "/assets/rfu/ui/brain_food.png"
+                "bloodshot" -> "/assets/rfu/ui/water.png"
                 "can_join" -> "/assets/rfu/ui/can_join.png"
                 "location" -> "/assets/rfu/ui/location.png"
                 "candy" -> "/assets/rfu/ui/candy.png"